python/storage_account/requirements.txt
azure-identity==1.17.1